Move HasGroups to separate module
