+++ pyiron_base/generic/hdfio.py
@@ -14,7 +14,7 @@ import h5io
 import numpy as np
 from tables.exceptions import NoSuchNodeError, HDF5ExtError
 import sys
-from pyiron_base.generic.has_groups import HasGroups
+from pyiron_base.interfaces.has_groups import HasGroups
 
 __author__ = "Joerg Neugebauer, Jan Janssen"
 __copyright__ = (
