+++ pyiron_base/job/core.py
@@ -10,7 +10,7 @@ import os
 import posixpath
 import math
 from pyiron_base.settings.generic import Settings
-from pyiron_base.generic.has_groups import HasGroups
+from pyiron_base.interfaces.has_groups import HasGroups
 from pyiron_base.generic.util import static_isinstance
 from pyiron_base.job.util import \
     _get_project_for_copy, \
